+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 - prefs: preferences now only show fonts that can render Chinese characters.
 - prefs: new advanced option to lock Chinese or English mode on CapsLock.
 
+### 🐛 Bug Fixes
+
+- tip: only handle keyboard_openclose change when in focus
+
 ### 🚜 Refactor
 
 - tip: fully model language state machine in the text input processor.